week_3_opea/OPEA-chatqna/requirements.txt
openai==1.12.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
tiktoken==0.6.0
tenacity==8.2.3
jinja2==3.1.4
aiofiles==23.2.1
